==> notebooks/2_split_normalization.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import os
from sklearn.preprocessing import StandardScaler
from iterstrat.ml_stratifiers import MultilabelStratifiedKFold
from sklearn.model_selection import train_test_split

model_data = pd.read_parquet('../data/model_data.parquet')

Time for the crucial part, splitting data. Images from the same patient should be in the same set. Also, the label distribution should be proportional. After doing some research I found a perfect package for this kind of job. Iterative-stratification, which will allow us to do it without unnecessarily complicated functions.

# First of all - let's group the data by 'Patient ID' and add all necessary columns. Let's assume that patients with multiple images have the state of their lungsconstant over time, 
# so we can use the first image as a representative of the patient - it may not rflect reality all the time, but it's adequate assumption in this case. We don't want to overcomplicate the splitting more than necessary.


# Group by 'Patient ID' and aggregate the data
grouped_data = (model_data.sort_values(['Patient ID', 'Follow-up #'])
                .groupby('Patient ID')
                .first()
                .reset_index())

print(grouped_data.head())


finding_vals = ['Atelectasis', 'Cardiomegaly', 'Effusion', 'Infiltration', 'Mass', 'Nodule', 'Pneumonia', 
                'Pneumothorax', 'Consolidation', 'Edema', 'Emphysema', 'Fibrosis', 'Pleural_Thickening', 'Hernia', 
                 'No Finding']



x = grouped_data['Patient ID']
y = grouped_data[finding_vals].values

MSKF = MultilabelStratifiedKFold(n_splits=5, shuffle=True, random_state=99)

train_idx, test_val_idx = next(MSKF.split(x, y))

train_patients = x.iloc[train_idx]
test_val_patients = x.iloc[test_val_idx]




# Now we can create a train/validation split from the training set. The same method will be used to ensure that the label distribution is proportional and that patients with multiple images are in the same set

x_test_val = grouped_data['Patient ID'].iloc[test_val_idx]
y_test_val = grouped_data[finding_vals].iloc[test_val_idx].values

MSKF_2 = MultilabelStratifiedKFold(n_splits=2, shuffle=True, random_state=99)

val_idx, test_idx = next(MSKF_2.split(x_test_val, y_test_val))

val_patients = x_test_val.iloc[val_idx]
test_patients = x_test_val.iloc[test_idx]




train_df = model_data[model_data['Patient ID'].isin(train_patients)].copy()
val_df = model_data[model_data['Patient ID'].isin(val_patients)].copy()
test_df = model_data[model_data['Patient ID'].isin(test_patients)].copy()

# Let's see if the MSKF worked as expected
def show_split_distribution(df):
    return df[finding_vals].mean()


ver_dict = {}
ver_dict['Train'] = show_split_distribution(train_df)
ver_dict['Validation'] = show_split_distribution(val_df)
ver_dict['Test'] = show_split_distribution(test_df)

print(ver_dict)

# We can see that the label distribution is relatively proportional in all sets - there are some differences in some categories, but it's expected, since they are rare

# Now let's remove Patient ID and Follow-up # columns, since we don't need them anymore. We will use 'image_id' as an unique ID for each sample
cols_to_remove = ['Patient ID', 'Follow-up #']

def remove_cols(df, cols):
    df = df.drop(columns=cols)
    return df

train_df = remove_cols(train_df, cols_to_remove)
val_df = remove_cols(val_df, cols_to_remove)
test_df = remove_cols(test_df, cols_to_remove)



Now it's time for normalization. We want to do on the train dataset, after splitting data - to prevent data leakage, we don't want the information from val/test set in our model during training


normalization_params = {"min_age": train_df['Patient Age'].min(),
                        "max_age": train_df['Patient Age'].max()}

train_df['PatientAge_norm'] = (train_df['Patient Age'] - normalization_params["min_age"]) / (normalization_params["max_age"] - normalization_params["min_age"])
train_df.drop(columns=['Patient Age'], inplace=True)


print(train_df.head())

val_df['PatientAge_norm'] = (val_df['Patient Age'] - normalization_params["min_age"]) / (normalization_params["max_age"] - normalization_params["min_age"])
val_df.drop(columns=['Patient Age'], inplace=True)

test_df['PatientAge_norm'] = (test_df['Patient Age'] - normalization_params["min_age"]) / (normalization_params["max_age"] - normalization_params["min_age"])
test_df.drop(columns=['Patient Age'], inplace=True)

Everything is ready, now we can export the datasets into separate files. <br>
In the next notebook we will work with images

train_df.to_parquet('../data/train_df.parquet')

val_df.to_parquet('../data/val_df.parquet')

test_df.to_parquet('../data/test_df.parquet')
