==> notebooks/3_image_preprocessing.ipynb <==
import os
import pandas as pd
from PIL import Image
from tqdm import tqdm




image_dir = '../data/versions/3/images'
output_base_dir = '../data/images'

train_df = pd.read_parquet('../data/train_df.parquet')
val_df = pd.read_parquet('../data/val_df.parquet')
test_df = pd.read_parquet('../data/test_df.parquet')

splits_dict = {
    'train': train_df,
    'val': val_df,
    'test': test_df
    }

target_size = (224, 224)



def resize_and_save_img(split_name, df):
    output_dir = os.path.join(output_base_dir, split_name)
    os.makedirs(output_dir, exist_ok=True)

    for image_id in tqdm(df['image_id'], desc=f"Processing {split_name}"):
        input_path = os.path.join(image_dir, image_id)
        output_path = os.path.join(output_dir, image_id)
        
        try:
            img = Image.open(input_path).convert('RGB')
            img = img.resize(target_size, Image.LANCZOS)
            img.save(output_path)
        except Exception as e:
            print(f"Failed to process {image_id}: {e}")


for split_name, df in splits_dict.items():
    resize_and_save_img(split_name, df)

print("Image preprocessing completed.")
